refactor: organize features with standard structure
