+++ src/features/ai/components/SessionInsights.tsx
