+++ src/features/notifications/index.ts
@@ -1,0 +1,1 @@
+export * from './components/NotificationBell'
