+++ src/features/patients/index.ts
@@ -1,0 +1,1 @@
+export * from './components/PatientList'
