+++ src/features/records/index.ts
@@ -1,0 +1,2 @@
+export * from './components/RecordForm'
+export * from './components/RecordList'
