+++ src/features/records/tests/recordForm.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { RecordForm } from '@/features/records/RecordForm';
+import { RecordForm } from '@/features/records';
 
 jest.mock('firebase/firestore', () => ({
   collection: () => ({}),
